# Conflicts:
#	HW05/test/ModelTest.java

+++ HW05/src/model/operation/color/OperationUtil.java
@@ -13,7 +13,8 @@ public class OperationUtil {
    * @param transformMatrix the matrix to check.
    * @throws IllegalArgumentException if the color transform matrix is not 3*3
    */
-  public static void checkColorTransformMatrix(double[][] transformMatrix) throws IllegalArgumentException {
+  public static void checkColorTransformMatrix(double[][] transformMatrix)
+          throws IllegalArgumentException {
     if (transformMatrix.length != 3 || transformMatrix[0].length != 3
             || transformMatrix[1].length != 3 || transformMatrix[2].length != 3) {
       throw new IllegalArgumentException("The Provided Color transformation matrix is invalid!");
@@ -25,7 +26,7 @@ public class OperationUtil {
    *
    * @param kernel the matrix to check.
    * @throws IllegalArgumentException filter kernel matrix is not a rectangular matrix with odd
-   * rows and columns.
+   *                                  rows and columns.
    */
   public static void checkKernel(double[][] kernel) throws IllegalArgumentException {
     boolean valid = false;
@@ -132,7 +133,7 @@ public class OperationUtil {
           }
         }
 
-        int[] rgb = OperationUtil.giveValidColorValue((int) r, (int) g, (int) b);
+        int[] rgb = OperationUtil.giveValidColorValue(r, g, b);
         filtered[i][j] = new Color(rgb[0], rgb[1], rgb[2], original[i][j].getAlpha());
       }
     }
